move everything back to vif
