+++ pkg/vif/buffer/data_darwin.go
