+++ pkg/vif/routing/router.go
@@ -1,0 +1,4 @@
+package routing
+
+type Router struct {
+}
