+++ pkg/vif/stack.go
@@ -1,4 +1,4 @@
-package netstack
+package vif
 
 import (
 	"context"
